requirements.txt
# Core dependencies
aiohttp>=3.9.0
asyncio>=3.4.3
pydantic>=2.5.0
click>=8.1.7
rich>=13.7.0
pyyaml>=6.0.1

# Web scraping and automation
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Authentication and security
keyring>=24.3.0
cryptography>=41.0.0

# Data processing
python-dateutil>=2.8.2
pandas>=2.1.0

# Export formats
jinja2>=3.1.2
markdown>=3.5.0

# Logging and monitoring
loguru>=0.7.2

# HTTP and requests
httpx>=0.25.0
urllib3>=2.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
